refactor(state): extract app state runtime into ts-state-tauri crate (#360)
Вынос рантайма состояния приложения из монолита `src-tauri/src/state/` (~18k LOC) в новый
крейт `ts-state-tauri`: StateManager/ProjectState/EventBus/CommandHandler/persistence +
browser/chat/workspace + commands/commands_api. Сюда же сложен `types_export.rs`
(specta-типы состояния — потреблялись только state).

ВАЖНО: монолитный `state/` НЕ использует существующий headless-крейт `ts-state` (#100) —
это ПАРАЛЛЕЛЬНЫЕ реализации ProjectState/EventBus. Монолитный слой tauri-связан (зависит от
ts-render-tauri/ts-media-tauri/ts-security), поэтому вынесен отдельным `ts-state-tauri`,
НЕ трогая `ts-state` #100. Реконсиляция дубля (ts-state core <- ts-state-tauri) — отдельно.

Пути: crate::video_compiler::commands::ai_api_proxy -> ts_render_tauri::ai_api_proxy;
crate::video_compiler::core -> ts_render; core::ffmpeg::analysis -> ts_render_analysis;
crate::security -> ts_security; crate::media::commands -> ts_media_tauri::commands;
crate::state:: (intra) -> crate::. Цикла нет (command-крейты не зависят от state).

Монолит: `state/`+`types_export.rs` -> файл-шимы `src/state.rs`/`src/types_export.rs`.
Зависимости: ts-render(+services/tauri/analysis), ts-media-tauri, ts-security, ts-core,
ts-schema, ffmpeg-next, reqwest, image, ... Разблокирует recognition/analysis/mcp.

Проверки: cargo test -p ts-state-tauri --test-threads=2 -> 9 passed;
cargo check --lib (src-tauri) -> 0 ошибок, 3 baseline-варнинга.

Closes #360. Часть #354.
