+++ crates/ts-state-tauri/src/browser.rs
