+++ crates/ts-state-tauri/src/commands/advanced_edits.rs
@@ -1,5 +1,5 @@
 use super::types::CommandResult;
-use crate::state::{EventBus, ProjectEvent, ProjectState};
+use crate::{EventBus, ProjectEvent, ProjectState};
 use std::sync::Arc;
 use tokio::sync::RwLock;
 
@@ -255,7 +255,7 @@ impl AdvancedEditsCommands {
       .publish(
         ProjectEvent::ClipUpdated {
           clip_id: clip_id.clone(),
-          changes: crate::state::events::ClipChanges {
+          changes: crate::events::ClipChanges {
             name: None,
             playback_rate: None,
             volume: None,
@@ -273,7 +273,7 @@ impl AdvancedEditsCommands {
       .publish(
         ProjectEvent::ClipUpdated {
           clip_id: adjacent_clip_id.clone(),
-          changes: crate::state::events::ClipChanges {
+          changes: crate::events::ClipChanges {
             name: None,
             playback_rate: None,
             volume: None,
@@ -363,7 +363,7 @@ impl AdvancedEditsCommands {
       .publish(
         ProjectEvent::ClipUpdated {
           clip_id: clip_id.clone(),
-          changes: crate::state::events::ClipChanges {
+          changes: crate::events::ClipChanges {
             name: None,
             playback_rate: None,
             volume: None,
@@ -562,7 +562,7 @@ impl AdvancedEditsCommands {
           .publish(
             ProjectEvent::ClipUpdated {
               clip_id: updated_id.clone(),
-              changes: crate::state::events::ClipChanges {
+              changes: crate::events::ClipChanges {
                 name: None,
                 playback_rate: None,
                 volume: None,
