+++ crates/ts-state-tauri/src/commands/color_grading.rs
@@ -1,6 +1,6 @@
 use super::types::CommandResult;
-use crate::state::project_state::*;
-use crate::state::{EventBus, ProjectEvent, ProjectState};
+use crate::project_state::*;
+use crate::{EventBus, ProjectEvent, ProjectState};
 use chrono::Utc;
 use std::sync::Arc;
 use tokio::sync::RwLock;
