+++ crates/ts-state-tauri/src/commands/effects.rs
@@ -1,5 +1,5 @@
 use super::types::CommandResult;
-use crate::state::{EventBus, ProjectEvent, ProjectState};
+use crate::{EventBus, ProjectEvent, ProjectState};
 use std::sync::Arc;
 use tokio::sync::RwLock;
 
@@ -60,7 +60,7 @@ impl EffectsCommands {
       .publish(
         ProjectEvent::ClipUpdated {
           clip_id: clip_id.clone(),
-          changes: crate::state::events::ClipChanges {
+          changes: crate::events::ClipChanges {
             name: None,
             playback_rate: None,
             volume: None,
@@ -126,7 +126,7 @@ impl EffectsCommands {
       .publish(
         ProjectEvent::ClipUpdated {
           clip_id: clip_id.clone(),
-          changes: crate::state::events::ClipChanges {
+          changes: crate::events::ClipChanges {
             name: None,
             playback_rate: None,
             volume: None,
